cmd: merged susecloud into the main binary
